proportion changed to Double

+++ Sources/SwiftGameOfLife/Grid.swift
@@ -14,7 +14,7 @@ public class Grid{
     public var cells: [[Cell]]
     public var rules: [Rule] = []
     
-    public init(width:Int = 10, height:Int = 10, isRandom:Bool = false, proportion:Int = 5) {
+    public init(width:Int = 10, height:Int = 10, isRandom:Bool = false, proportion:Double = 5.0) {
         self.width = width
         self.height = height
         self.cells = []
@@ -37,12 +37,12 @@ public class Grid{
         }
     }
     
-    public func initRandomGrid(_ width: Int, _ height: Int, _ proportion:Int = 5) {
+    public func initRandomGrid(_ width: Int, _ height: Int, _ proportion:Double = 5.0) {
         cells = []
         for i in 0..<width{
             var row:[Cell] = []
             for j in 0..<height{
-                let state:CellState = (Int.random(in: 0 ..< 10)>proportion) ? .alive : .dead
+                let state:CellState = (Double.random(in: 0 ..< 10)>proportion) ? .alive : .dead
                 row.append(Cell(x: i, y: j , state: state))
             }
             cells.append(row)
